# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GraphRAG Orchestrator",
    "version": "1.0.0"
  },
  "paths": {
...
    "/ingest": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Submit documents for ingestion",
        "description": "Submit base64-encoded documents for entity extraction and knowledge graph ingestion. Supports sync and async modes.",
        "operationId": "ingest_ingest_post",
        "parameters": [
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Sync"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ingest/{job_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Check ingestion job status",
        "description": "Retrieve the current status and result of an async ingestion job.",
        "operationId": "get_ingest_job_ingest__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Execute a graph query",
        "description": "Submit a natural-language query against the knowledge graph. Supports sync and async modes with automatic complexity routing.",
        "operationId": "query_query_post",
        "parameters": [
          {
            "name": "async_mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Async Mode"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/{job_id}": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Check query job status",
        "description": "Retrieve the current status and result of an async query job.",
        "operationId": "get_query_job_query__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/query/{query_id}/evaluation": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Get query evaluation",
        "description": "Retrieve the RAG evaluation result for a completed query.",
        "operationId": "get_evaluation_query__query_id__evaluation_get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestDocument": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "repository": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository"
          },
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "content",
          "source_type"
        ],
        "title": "IngestDocument"
      },
      "IngestJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at"
        ],
        "title": "IngestJobResponse"
      },
      "IngestRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestDocument"
            },
            "type": "array",
            "minItems": 1,
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "IngestRequest"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
      "QueryComplexity": {
        "type": "string",
        "enum": [
          "entity_lookup",
          "single_hop",
          "multi_hop",
          "aggregate"
        ],
        "title": "QueryComplexity"
      },
      "QueryJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "number",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at"
        ],
        "title": "QueryJobResponse"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "minLength": 1,
            "title": "Query"
          },
          "max_results": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Results",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources"
          },
          "complexity": {
            "$ref": "#/components/schemas/QueryComplexity"
          },
          "retrieval_path": {
            "type": "string",
            "title": "Retrieval Path"
          },
          "evaluation_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Score"
          },
          "retrieval_quality": {
            "type": "string",
            "title": "Retrieval Quality",
            "default": "skipped"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "complexity",
          "retrieval_path"
        ],
        "title": "QueryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}